Add files via upload
This project is ultimately meant to generate a random and unique word that follows the phonotactic restraints of the English language.
Current status:
"Success! I now have a randomly generated word with a random number of syllables and varying syllable length!"

Next step is to introduce phonotactic constraints so that the word reads and sounds naturally like a word in English.

Eventually, I intend to adapt this program to generate unique names of a fantasy style.